requirements.txt
Django==4.2.26
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.1
python-decouple==3.8
psycopg2-binary==2.9.11
Pillow==11.3.0

